backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
faiss-cpu
openai
python-multipart
